move store and GlobalStyles out of the components dir
