+++ react-app-mui/src/components/items/logic/ItemFunc.ts
@@ -2,9 +2,9 @@ import {useNavigate} from "react-router-dom";
 import dateFormat from "dateformat";
 import {useDispatch, useSelector} from "react-redux";
 import {ITask} from "../../../helpers/interfaces";
-import {changeItemImportant, deleteItem, filterItems} from "../../store/itemsSlice";
+import {changeItemImportant, deleteItem, filterItems} from "../../../store/itemsSlice";
 import {useCallback, useEffect} from "react";
-import {selectFilter} from "../../store/store";
+import {selectFilter} from "../../../store/store";
 
 
 const ItemFunc = (item: ITask) => {
