+++ react-app-mui/src/components/items/logic/ItemsFunc.tsx
@@ -1,12 +1,12 @@
-import {selectFilter, selectItems} from "../../store/store";
+import {selectFilter, selectItems} from "../../../store/store";
 import {useDispatch, useSelector} from "react-redux";
 import {FilterOption} from "../styles/Items/FilterContainer";
 import {useOnClickOutside} from "usehooks-ts";
 import {useEffect, useLayoutEffect, useRef} from "react";
 import {IOptions} from "../../../helpers/interfaces";
-import {changeOption, closeFilterDropdown, toggleFilterDropdown} from "../../store/filterSlice";
+import {changeOption, closeFilterDropdown, toggleFilterDropdown} from "../../../store/filterSlice";
 import {useLocation} from "react-router-dom";
-import {filterItems, getAllItems} from "../../store/itemsSlice";
+import {filterItems, getAllItems} from "../../../store/itemsSlice";
 
 
 const options: IOptions[] = ["A-Z", "Z-A", "Ważne", "Najwcześniejsza Data", "Najpóźniejsza Data"];
