+++ react-app-mui/src/components/navbar/logic/NavbarFunc/NavbarFunc.ts
@@ -1,6 +1,6 @@
 import {useDispatch} from "react-redux";
 import {useNavigate} from "react-router-dom";
-import {changeOption} from "../../../store/filterSlice";
+import {changeOption} from "../../../../store/filterSlice";
 
 
 const NavbarFunc = () => {
